pyproject.toml
[tool.poetry]
name = "django-redis-blog"
version = "0.1.0"
description = "Simple Blog API created with Django and Redis as Caching Backend"
authors = ["Shakib Hossain <shakib609@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
django-redis-cache = "^2.1"
drf-extensions = "^0.5.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
